supports members and aliasing when degraded
