+++ src/index.js
@@ -13,13 +13,14 @@ if (typeof window !== 'undefined') {
             Container = __hilary.Container(locale, polyn.is, polyn.Immutable, Exception),
             Context = __hilary.Context(polyn.Immutable, Container),
             HilaryModule = __hilary.HilaryModule(polyn.is, polyn.Blueprint, polyn.objectHelper, locale, Exception),
-            Logger = __hilary.Logger(polyn.is);
+            Logger = __hilary.Logger(polyn.is),
+            ResolveTasks = __hilary.ResolveTasks(polyn.async, polyn.is, locale, Exception);
 
             polyn.Blueprint.configure({
                 compatibility: '2017-03-20'
             });
 
-        window.hilary = __hilary.HilaryApi(polyn.async, polyn.is, polyn.id, polyn.Immutable, locale, Logger, Exception, Context, HilaryModule);
+        window.hilary = __hilary.HilaryApi(polyn.async, polyn.is, polyn.id, polyn.Immutable, locale, Logger, Exception, Context, HilaryModule, ResolveTasks);
 
     }(window.polyn, window.__hilary, window));
 } else if (typeof module !== 'undefined' && module.exports) {
@@ -30,7 +31,8 @@ if (typeof window !== 'undefined') {
         Context = require('./Context')(polyn.Blueprint, Container, Exception, locale),
         HilaryModule = require('./HilaryModule')(polyn.is, polyn.Blueprint, polyn.objectHelper, locale, Exception),
         Logger = require('./Logger')(polyn.is),
-        hilary = require('./HilaryApi')(polyn.async, polyn.is, polyn.id, polyn.Immutable, locale, Logger, Exception, Context, HilaryModule);
+        ResolveTasks = require('./ResolveTasks')(polyn.async, polyn.is, locale, Exception),
+        hilary = require('./HilaryApi')(polyn.async, polyn.is, polyn.id, polyn.Immutable, locale, Logger, Exception, Context, HilaryModule, ResolveTasks);
 
         polyn.Blueprint.configure({
             compatibility: '2017-03-20'
